requirements.txt
python-telegram-bot
pytest
flask
waitress

pytest-cov
codecov